Fixes broken obs detail with obscured coords (closes #145)
Note that the API should *never* give the user private_latitude and
private_longitude if they don't have permission to view them, so it's safe to
display them in the client regardless of the geoprivacy. Also note that the
observation geoprivacy is not the only determinant of coordinate obscuration.
Coordinates can be obscured because the taxon is threatened too.