Partial fix for observation list item text crop
This removes an unneeded viewgroup and improves the listitem slightly. By changing the item height
to wrap_content, we allow the row itself to grow slightly if user asks for a large font size. To avoid
being ridiculous, we align the top and bottom of the location text to the location_icon. This will not work
on all devices, because some scale text much faster than images, so we can still see cropping if the
text is scaled up to 20dp but the location icon is not scaled up. However, it improves the situation on
standard android and increases layout performance slightly. The real proper fix is to use a constraintlayout