obs list: force iconic pic to match layout of pic
By setting constraints on all four sides and setting the width/height to match constraint (e.g. 0dp),
the iconic pic will always be sized and aligned identically to the regular observation pic. This is easier
than always trying to keep their XML in sync, and it make it easier to position other UI elements relative
to the "picture" - we always position them relative to the observation picture, and the iconic picture will
follow the observation picture

This allows us to remove the barrier element as well as better indicates the purpose of the taxonic picture -
it is shown instead of the observation picture if the obs pic has not yet loaded. Not quite sure why we do
not just use one ImageView, but I'm keeping it consistent until I grok the code