obs list: Better prevention of view overlap
In constrained horizontal space, the species guess (or location guess) will overlap the
content on the right of the list. We already had a solution for this in place, but in the case
of potential overlap it would add ellipsis to the text without any margin between the
ellipses and the item we are trying to not overlap. By adding a simple margin_right we
get the same non-overlap with the added benefit of preserving at least some margin
if we have to shorten the species or location guess textview.

Additionally, this prevents the case of species guess overlapping the upload progress bar
by adding a barrier. Admittedly it's a very small edge-case, but it costs us nothing to
handle it properly