obs list: Ensure margin when in multi-select mode
This adds a marginLeft between the checkbox and the observation pic. We add margin ot the
pic instead of a marginRight to the checkbox b/c the constraintLayout only respects margins
between your view and the views you are constrained by. Pic is constrained by checkbox, not
the other way around, so pic has the margin.

Subtle note - the checkbox is Visibility.GONE when not in use. We have to explicitely tell the
constraintlayout to ignore the margin in that scenario (as opposed to adding a margin to the
size 0,0 checkbox location)