obs list: allow location icon to be GONE
Previously if location_icon was marked GONE (as it is during upload or errors),
the layout would all get broken. Now we just add a goneMarginTop that equals the
explicitely-set height of the icon. If the icon is GONE, the layout should still be
identical and the place_guess text will shift left as desired.

Also fix a minor bug - if id_count is not displayed b/c there are zero, we still want
the comment count to be baseline-aligned with place_guess, so we align it
directly instead of aligning with the id_count which was alignted with place_guess