Fix bug - constraint layout groups lock vis setting
Thought we could use a group to set visibility on all sub-views at
once. While this works, it breaks ability to set vis on each subview manually.
All subviews in the group are locked to the same visibility at all times.
Workaround is to build a small custom viewgroup. Hate this, because we are
allocating a new view inside the getView call so it's not so good performance-wise
but it makes the code work and my metrics show this tiny memory allocation is
still much better than the prior (4-5 nested relative/linear layout approach)