obs list: Use constraint barrier for place guess
Previously we were constraining to the comment_pic. However, this view is changed
by the cursoradapter quite a bit. Sometimes it's INVISIBLE, someitmes GONE. Instead,
we use a proper barrier so we never have to worry about which of these 5 views are
in/visible/gone, we just barrier to the start value. New benefit is that we no longer
overlap the progress indicator (it was a small edge case)