obs list: Mark unused Views GONE to expand layout
Using invisible hides the element but does not allow the other layout
views to expand into that space. By using gone instead, we properly
allow using all available space.

Also adjusted the margin End if either of our barriers become GONE
(this happens when all elements referenced by barrier are GONE). This
actually happens a lot, because we use the commentIdContainer in the
adapter to GONE the id/comment image+text all at once. In this case,
we no longer need the 8dp margin and we set it to 0dp and rely on
the padding already present in the parent constraintlayout