Fix Observation tab rendering in AS layout viewer
No functional changes. Fixes the tab divider to properly match the height of the
LinearLayout (it will always match in reality because they have the same parent).
We *could* have just set the linearlayout to match_parent, but that looks funny in the
editor (it fills entire 'screen' when looking at this one tab).

Also removes the 'gone' marker on the count. Code controls this attribute anyway,
so allowing it to show in layout viewer just makes understanding the layout files a
little bit simpler

Lastly fixes the text on observations to match the code (which uses CAPS)