Transition all Service HTTP calls to OkHttp
Now that I've got high code coverage in the legacy request() method,
I have been able to prove to myself that the new request() method works
as well or better.

The next step is to start pushing all the networking code inside the API
package, with the intention of eventually replacing 90% of that code with
either okhttp or retrofit. Before we can replace it, we have to cleanly
separate what's our own business logic and what's networking code. This
is the first big step in that direction.

All of the requests are now internally handled by okhttp. We've left the
get/post/reqest/delete helper methods in the service for now because
they are API incompatible for the moment (they return JSONArray while
the new API returns ApiResponse).