obs list: Add section tracing inside getView
This is one of the most active code paths in the application, so having precise timing
data on it is critical. Note that these calls are effectively a no-op unless tracing is enabled,
they are safe to leave in production. Leaving them in makes it simple to see this detailed data
anytime you trace the app for performance checks