obslist: Cache taxonic icons for performance
The old implementation used a large if-else. Java can specifically optimize a
large string switch statement, because it uses object.hashcode to make it a
hash lookup of time O(log(n)) vs a string lookup of O(n). After I did this, I saw
we were still spending time in setImageResource just decoding the bitmaps
directly on the UI thread. It's only a minor performance lag to the UI thread,
but 1) as we use these icons massively when scrolling this list, 2) they are tiny
in total size e.g. 10KB, and 3) every ms counts on the UI thread because we are
already dropping frames. All this considered, I decided to trade CPU time for
memory. We now lazily load the taxon icon images into memory, and hold them
in the static memory. THis somewhat defeats the switch because we only
actually run that code ~10 times now (once for each taxon icon), but I had already
written it and it's objectively faster so I left it in.

Overall these improvements only give a small performance bump, less than 0.5ms
on the UI thread. So nothing to write home about, but I will take it as that's 0.5ms
we can spend doing other stuff before dropping a frame (and it might be more than
0.5 on a slower device)