obslist: Improve performance of loadObsImage
This method runs on the UI thread, which means slowdowns here cause dropped frames. Currently
this happens a fair amount. Before this commit, this method takes about 4.5ms, and after this commit
it takes about 2ms on my test device.  This is in the fully online scenario, no offline was considered.

The main issue was identified to be the call to mApp.isLowMem, which internally runs a binder transaction.
This was fixed by moving the call from the front of the method, where it ran 100% of the time, to deeper
inside the method logic, so it only runs if all the other conditional flows fail. In the expected case, we will
still run this method once per call to loadObs, so if we could refactor it away further that would take us from
about 2ms to about 0ms. Note that loadObs method is recursive.

Secondary issue was formatting log directives. We solved this by delaying formatting until the log is insidethe
tinylog thread, by passing a format string and args instead of passing a pre-formatted string.

We also tweaked isLowMem to delay the binder call if it's not needed because the method can fast-fail. Unlikely
this will cause significant improvements in the average case, but it should help in low mem scenarios e.g. the
worst case