Compile smoothrescale using SDK=23
It's generally advisable to align the native minSdkVersion used for the application with the
APP_PLATFORM version that is targeted by the JNI layer. This ensures all optimizations are
available for the native code, but also ensures there is alignment of the runtime contract
used in the native layer with the runtime contract used in the Java layers. AFAIK there is
no negative to doing this, it's a compilation check mainly to ensure you do not use older
data, but doing so will also guarantee you have API contract alignment