Test SDK 30, not 33
Upgraded the androidx.test.core to address this issue: https://github.com/android/android-test/issues/1412

Still having issues with flaky tests on SDK 33. Appears the way that focus is handled in
the application is problematic for espresso. A quick look turned up some non-obvious code
(the code under test handles clicks but returns false indicating that it did not handle
them, and the view is a nest of activity-fragment-fragment that will take me a moment to
unpack).

Chose to just maintain the SDK that was being tested previosuly (30) but put 33 into the
list as something to be fixed