Store CI artifacts; upgrade tooling; wait on boot
This contains multiple improvements to the CI/CD process. Main one is giving the
emulator much, much longer to finish booting. Emulators are fickle -
the sys.boot_completed property indicates the init system is ready, it DOES NOT indicate
that the emulator is ready for testing. You can see this in the screenshots - simply waiting
shows you the emulator launcher continuing to boot, the clock being set properly, etc.

This also stores multiple artifacts. Notably the emulator build reports and any screenshots
we have caputured. I've also setup to store the APKs - not really needed just yet since
iNat is not using the CI/CD to build authoritative images, but nice to have so you can
easily install a certain debug build without needing to build it locally. If iNat parent
project chooses, they could update the CI to inject proper secrets into the APK so the
debug APKs are usable.

This also runs compilation before testing to avoid mixing timing for compile and test.
Compilation is almost never the problem, it's always the instrumented tests and particularly
the emulator that is giving issues.

Lastly, this ensures the matrix will continue to run even if one other part of the matrix
fails. This is important as we work towards adoption of 33 and 34.